Move Types out of `pages` dir to keep Next.js happy
Having non-pages under pages/ breaks the Next.js build.
